Work on saving chats in redis

+++ internal/models/chat_record.go
@@ -3,7 +3,7 @@ package models
 // Represents a single back-and-forth chat record
 // Will be used to represent a chat record in both Redis and SurrealDB
 type ChatRecord struct {
-	ChatId int    `json:"chatId"`
+	ChatId int64  `json:"chatId"`
 	Prompt string `json:"prompt"`
 	Reply  string `json:"reply"`
 }
